# PED_SERVER_API/src/IO.Swagger/Models/State.cs
/*
 * PED Test API
 *
 * PED Test API
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class State : IEquatable<State>
    { 
        /// <summary>
        /// Gets or Sets StateId
        /// </summary>

        [DataMember(Name="stateId")]
        public decimal? StateId { get; set; }

        /// <summary>
        /// Gets or Sets StateName
        /// </summary>

        [DataMember(Name="stateName")]
        public string StateName { get; set; }

        /// <summary>
        /// Gets or Sets StateCode
        /// </summary>

        [DataMember(Name="stateCode")]
        public string StateCode { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class State {\n");
            sb.Append("  StateId: ").Append(StateId).Append("\n");
            sb.Append("  StateName: ").Append(StateName).Append("\n");
            sb.Append("  StateCode: ").Append(StateCode).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((State)obj);
        }

        /// <summary>
        /// Returns true if State instances are equal
        /// </summary>
        /// <param name="other">Instance of State to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(State other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    StateId == other.StateId ||
                    StateId != null &&
                    StateId.Equals(other.StateId)
                ) && 
                (
                    StateName == other.StateName ||
                    StateName != null &&
                    StateName.Equals(other.StateName)
                ) && 
                (
                    StateCode == other.StateCode ||
                    StateCode != null &&
                    StateCode.Equals(other.StateCode)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (StateId != null)
                    hashCode = hashCode * 59 + StateId.GetHashCode();
                    if (StateName != null)
                    hashCode = hashCode * 59 + StateName.GetHashCode();
                    if (StateCode != null)
                    hashCode = hashCode * 59 + StateCode.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(State left, State right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(State left, State right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}