# PED_SERVER_API/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "PED Test API",
    "description" : "PED Test API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/SSDMUKK/Swagger_PED/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  }, {
    "url" : "https://virtserver.swaggerhub.com/DEVELOPMENTSUPPORT_1/PED/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  } ],
  "tags" : [ {
    "name" : "admin",
    "description" : "Secured Admin-only calls"
  }, {
    "name" : "developers",
    "description" : "Operations available to regular developers"
  }, {
    "name" : "shared",
    "description" : "Shared calls"
  } ],
  "paths" : {
    "/inventory" : {
      "get" : {
        "tags" : [ "admin" ],
        "summary" : "searches inventory",
        "description" : "By passing in the appropriate options, you can search for\navailable inventory in the system\n",
        "operationId" : "searchInventory",
        "parameters" : [ {
          "name" : "searchString",
          "in" : "query",
          "description" : "pass an optional search string for looking up inventory",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "skip",
          "in" : "query",
          "description" : "number of records to skip for pagination",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "maximum number of records to return",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "maximum" : 50,
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/InventoryItem"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          }
        }
      },
      "post" : {
        "tags" : [ "admin" ],
        "summary" : "adds an inventory item",
        "description" : "Adds an item to the system",
        "operationId" : "addInventory",
        "requestBody" : {
          "description" : "Inventory item to add",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/InventoryItem"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "item created"
          },
          "400" : {
            "description" : "invalid input, object invalid"
          },
          "409" : {
            "description" : "an existing item already exists"
          }
        }
      }
    },
    "/states" : {
      "get" : {
        "tags" : [ "shared" ],
        "summary" : "Get all States",
        "operationId" : "states",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/State"
                  },
                  "x-content-type" : "application/json"
                },
                "example" : [ {
                  "stateId" : 1,
                  "stateName" : "LC"
                }, {
                  "stateId" : 2,
                  "stateName" : "NC"
                } ]
              }
            }
          }
        }
      }
    },
    "/reasoncodes" : {
      "get" : {
        "tags" : [ "shared" ],
        "summary" : "Get all Reason Codes",
        "operationId" : "reason codes",
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ReasonCode"
                  },
                  "x-content-type" : "application/json"
                },
                "example" : [ {
                  "reasonCodeId" : 1,
                  "reasonCodeName" : "LC",
                  "groupId" : 1
                }, {
                  "reasonCodeId" : 2,
                  "reasonCodeName" : "NC",
                  "groupId" : 1
                } ]
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "InventoryItem" : {
        "required" : [ "id", "manufacturer", "name", "releaseDate" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
          },
          "name" : {
            "type" : "string",
            "example" : "Widget Adapter"
          },
          "releaseDate" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2016-08-29T09:12:33.001Z"
          },
          "manufacturer" : {
            "$ref" : "#/components/schemas/Manufacturer"
          }
        },
        "example" : {
          "releaseDate" : "2016-08-29T09:12:33.001Z",
          "name" : "Widget Adapter",
          "id" : "d290f1ee-6c54-4b01-90e6-d701748f0851",
          "manufacturer" : {
            "phone" : "408-867-5309",
            "name" : "ACME Corporation",
            "homePage" : "https://www.acme-corp.com"
          }
        }
      },
      "Manufacturer" : {
        "required" : [ "name" ],
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "example" : "ACME Corporation"
          },
          "homePage" : {
            "type" : "string",
            "format" : "url",
            "example" : "https://www.acme-corp.com"
          },
          "phone" : {
            "type" : "string",
            "example" : "408-867-5309"
          }
        },
        "example" : {
          "phone" : "408-867-5309",
          "name" : "ACME Corporation",
          "homePage" : "https://www.acme-corp.com"
        }
      },
      "State" : {
        "type" : "object",
        "properties" : {
          "stateId" : {
            "type" : "number"
          },
          "stateName" : {
            "type" : "string"
          },
          "stateCode" : {
            "type" : "string"
          }
        },
        "example" : {
          "stateName" : "stateName",
          "stateId" : 0.8008281904610115,
          "stateCode" : "stateCode"
        }
      },
      "ReasonCode" : {
        "type" : "object",
        "properties" : {
          "reasonCodeId" : {
            "type" : "integer"
          },
          "reasonCodeName" : {
            "type" : "string"
          },
          "groupId" : {
            "type" : "integer"
          }
        },
        "example" : {
          "reasonCodeName" : "reasonCodeName",
          "groupId" : 6,
          "reasonCodeId" : 0
        }
      }
    }
  },
  "x-azure-api-id" : "sh-d896d186-d5c4-4a04-b584-12d584d0cbf5"
}